Bugfix and functionality update + cleanup of menusystem. Added Falagard derivatives of DefaultWindow, DragContainer and ItemEntry with minimal StateImagery. Removed TextItem as falagard now handles that exclusively.

+++ src/elements/CEGUIItemListBase.cpp
@@ -45,7 +45,7 @@ ItemListBaseProperties::AutoResizeEnabled	ItemListBase::d_autoResizeEnabledPrope
 *************************************************************************/
 // event names
 const String ItemListBase::EventNamespace("ItemListBase");
-const String ItemListBase::EventListContentsChanged( "ListItemsChanged" );
+const String ItemListBase::EventListContentsChanged("ListItemsChanged");
 
 	
 /*************************************************************************
@@ -69,15 +69,6 @@ ItemListBase::ItemListBase(const String& type, const String& name)
 ItemListBase::~ItemListBase(void)
 {
 	resetList_impl();
-}
-
-
-/*************************************************************************
-	Initialise the Window based object ready for use.
-*************************************************************************/
-void ItemListBase::initialise(void)
-{
-	//layoutItemWidgets();
 }
 
 
@@ -153,7 +144,6 @@ bool ItemListBase::isItemInList(const ItemEntry* item) const
 }
 
 
-
 /*************************************************************************
 	Remove all items from the list.
 *************************************************************************/
@@ -161,8 +151,7 @@ void ItemListBase::resetList(void)
 {
 	if (resetList_impl())
 	{
-		WindowEventArgs args(this);
-		onListContentsChanged(args);
+		handleUpdatedItemData();
 	}
 
 }
@@ -178,8 +167,7 @@ void ItemListBase::addItem(ItemEntry* item)
 		// just stick it on the end.
 		d_listItems.push_back(item);
 		addChildWindow(item);
-		WindowEventArgs args(this);
-		onListContentsChanged(args);
+		handleUpdatedItemData();
 	}
 
 }
@@ -215,8 +203,7 @@ void ItemListBase::insertItem(ItemEntry* item, const ItemEntry* position)
 		d_listItems.insert(ins_pos, item);
 		addChildWindow(item);
 
-		WindowEventArgs args(this);
-		onListContentsChanged(args);
+		handleUpdatedItemData();
 	}
 
 }
@@ -227,26 +214,16 @@ void ItemListBase::insertItem(ItemEntry* item, const ItemEntry* position)
 *************************************************************************/
 void ItemListBase::removeItem(ItemEntry* item)
 {
-	if (item)
+	if (item && isItemInList(item))
 	{
-		ItemEntryList::iterator pos = std::find(d_listItems.begin(), d_listItems.end(), item);
-
-		// if item is in the list
-		if (pos != d_listItems.end())
-		{
-			// remove item
-			d_listItems.erase(pos);
-
-			removeChildWindow(item);
-			WindowManager::getSingleton().destroyWindow(item);
-
-			WindowEventArgs args(this);
-			onListContentsChanged(args);
-		}
-
+	    removeChildWindow(item);
+	    if (item->isDestroyedByParent())
+	    {
+	        WindowManager::getSingleton().destroyWindow(item);
+	    }
 	}
-	
 }
+
 
 /*************************************************************************
 	Set wheter or not this ItemListBase widget should automatically
@@ -257,13 +234,14 @@ void ItemListBase::setAutoResizeEnabled(bool setting)
 	bool old = d_autoResize;
 	d_autoResize = setting;
 
-	// if not already enabled, trigger a resize
-	if ( d_autoResize && !old )
+	// if not already enabled, trigger a resize - only if not currently initialising
+	if ( d_autoResize && !old && !d_initialising)
 	{
 		sizeToContent();
 	}
 
 }
+
 
 /*************************************************************************
 	Causes the list box to update it's internal state after changes have
@@ -290,30 +268,42 @@ void ItemListBase::addItemListBaseEvents(void)
 *************************************************************************/
 void ItemListBase::onListContentsChanged(WindowEventArgs& e)
 {	
-	requestRedraw();
-
-	// if auto resize is enabled - do it
-	if ( d_autoResize )
+    // if we are not currently initialising we might have things todo
+	if (!d_initialising)
 	{
-		sizeToContent();
+	    requestRedraw();
+	    
+	    // if auto resize is enabled - do it
+	    if (d_autoResize)
+	    {
+		    sizeToContent();
+	    }
+	    // redo the item layout and fire our event
+	    layoutItemWidgets();
+	    fireEvent(EventListContentsChanged, e, EventNamespace);
 	}
-	
-	layoutItemWidgets();
-	fireEvent(EventListContentsChanged, e, EventNamespace);
 }
 
 
-/*************************************************************************
-	Handler for when we are sized
+/************************************************************************
+    Handler for when a child is removed
 *************************************************************************/
-void ItemListBase::onSized(WindowEventArgs& e)
+void ItemListBase::onChildRemoved(WindowEventArgs& e)
 {
-	// base class handling
-	Window::onSized(e);
+    // make sure it is removed from the itemlist if we have an ItemEntry
+    if (e.window->testClassName("ItemEntry"))
+    {
+        ItemEntryList::iterator pos = std::find(d_listItems.begin(), d_listItems.end(), e.window);
 
-	//layoutItemWidgets();
-
-	e.handled = true;
+        // if item is in the list
+        if (pos != d_listItems.end())
+        {
+            // remove item
+            d_listItems.erase(pos);
+            // trigger list update
+            handleUpdatedItemData();
+		}
+    }
 }
 
 
@@ -330,11 +320,14 @@ bool ItemListBase::resetList_impl(void)
 	// we have items to be removed and possible deleted
 	else
 	{
-		// delete any items we are supposed to ---------- DESTROYS ALL ITEMS !!!
+		// delete any items we are supposed to
 		for (size_t i = 0; i < getItemCount(); ++i)
 		{
 			removeChildWindow(d_listItems[i]);
-			WindowManager::getSingleton().destroyWindow(d_listItems[i]);
+			if (d_listItems[i]->isDestroyedByParent())
+			{
+			    WindowManager::getSingleton().destroyWindow(d_listItems[i]);
+			}
 		}
 
 		// clear out the list.
@@ -367,10 +360,37 @@ void ItemListBase::addChild_impl(Window* wnd)
 	{
 		// perform normal addItem - please no more infinite recursion !#?¤
 		d_listItems.push_back((ItemEntry*)wnd);
-		WindowEventArgs args(this);
-		onListContentsChanged(args);
+		handleUpdatedItemData();
 	}
 
+}
+
+
+/************************************************************************
+    Initialisation done
+*************************************************************************/
+void ItemListBase::endInitialisation(void)
+{
+    Window::endInitialisation();
+    handleUpdatedItemData();
+}
+
+
+/************************************************************************
+	Perform child window layout
+************************************************************************/
+void ItemListBase::performChildWindowLayout(void)
+{
+	Window::performChildWindowLayout();
+	// if we are not currently initialising
+	if (!d_initialising)
+	{
+	    // Redo the item layout.
+	    // We don't just call handleUpdateItemData, as that could trigger a resize,
+	    // which is not what is being requested.
+	    // It would also cause infinite recursion... so lets just avoid that :)
+	    layoutItemWidgets();
+	}
 }
 
 } // End of  CEGUI namespace section
